+++ v2ray.sh
@@ -26,6 +26,7 @@ https://www.23xsw.cc/
 )
 
 CONFIG_FILE="/etc/v2ray/config.json"
+SERVICE_FILE="/etc/systemd/system/v2ray.service"
 OS=`hostnamectl | grep -i system | cut -d: -f2`
 
 V6_PROXY=""
@@ -33,6 +34,16 @@ IP=`curl -sL -4 ip.sb`
 if [[ "$?" != "0" ]]; then
     IP=`curl -sL -6 ip.sb`
     V6_PROXY="https://gh.hijk.art/"
+fi
+
+BT="false"
+NGINX_CONF_PATH="/etc/nginx/conf.d/"
+START_NGINX="systemctl start nginx"
+res=`which bt`
+if [[ "$res" != "" ]]; then
+    BT="true"
+    NGINX_CONF_PATH="/www/server/panel/vhost/nginx/"
+    START_NGINX="nginx -c /www/server/nginx/conf/nginx.conf"
 fi
 
 VLESS="false"
@@ -421,17 +432,26 @@ getData() {
 
 installNginx() {
     colorEcho $BLUE " 安装nginx..."
-    if [[ "$PMT" = "yum" ]]; then
-        $CMD_INSTALL epel-release 
+    if [[ "$BT" = "false" ]]; then
+        if [[ "$PMT" = "yum" ]]; then
+            $CMD_INSTALL epel-release 
+        fi
+        $CMD_INSTALL nginx
+        systemctl enable nginx
+    else
+        res=`which nginx`
+        if [[ "$?" != "0" ]]; then
+            colorEcho $RED " 您安装了宝塔，请在宝塔后台安装nginx后再运行本脚本"
+            exit 1
+        fi
     fi
-    $CMD_INSTALL nginx
-    systemctl enable nginx
 }
 
 getCert() {
     if [[ -z ${CERT_FILE+x} ]]; then
-        systemctl stop nginx
         systemctl stop v2ray
+        nginx -s stop
+        sleep 2
         res=`netstat -ntlp| grep -E ':80|:443'`
         if [[ "${res}" != "" ]]; then
             colorEcho ${RED}  " 其他进程占用了80或443端口，请先关闭再运行一键脚本"
@@ -478,6 +498,10 @@ getCert() {
 
         CERT_FILE="/etc/letsencrypt/live/${DOMAIN}/fullchain.pem"
         KEY_FILE="/etc/letsencrypt/live/${DOMAIN}/privkey.pem"
+
+        sed -i '/certbot/d' /etc/crontab
+        certbotpath=`which certbot`
+        echo "0 3 1 */2 0 root nginx -s stop; ${certbotpath} renew ; $START_NGINX" >> /etc/crontab
     else
         mkdir -p /etc/v2ray
         cp ~/v2ray.pem /etc/v2ray/${DOMAIN}.pem
@@ -486,21 +510,23 @@ getCert() {
 }
 
 configNginx() {
-    if [[ ! -f /etc/nginx/nginx.conf.bak ]]; then
-        mv /etc/nginx/nginx.conf /etc/nginx/nginx.conf.bak
-    fi
-    res=`id nginx`
-    if [[ "$?" != "0" ]]; then
-        user="www-data"
-    else
-        user="nginx"
-    fi
     mkdir -p /usr/share/nginx/html;
     if [[ "$ALLOW_SPIDER" = "n" ]]; then
         echo 'User-Agent: *' > /usr/share/nginx/html/robots.txt
         echo 'Disallow: /' >> /usr/share/nginx/html/robots.txt
     fi
-    cat > /etc/nginx/nginx.conf<<-EOF
+
+    if [[ "$BT" = "false" ]]; then
+        if [[ ! -f /etc/nginx/nginx.conf.bak ]]; then
+            mv /etc/nginx/nginx.conf /etc/nginx/nginx.conf.bak
+        fi
+        res=`id nginx`
+        if [[ "$?" != "0" ]]; then
+            user="www-data"
+        else
+            user="nginx"
+        fi
+        cat > /etc/nginx/nginx.conf<<-EOF
 user $user;
 worker_processes auto;
 error_log /var/log/nginx/error.log;
@@ -537,6 +563,7 @@ http {
     include /etc/nginx/conf.d/*.conf;
 }
 EOF
+    fi
 
     if [[ "$PROXY_URL" = "" ]]; then
         action=""
@@ -549,11 +576,11 @@ EOF
     fi
 
     if [[ "$TLS" = "true" || "$XTLS" = "true" ]]; then
-        mkdir -p /etc/nginx/conf.d;
+        mkdir -p $NGINX_CONF_PATH
         # VMESS+WS+TLS
         # VLESS+WS+TLS
         if [[ "$WS" = "true" ]]; then
-            cat > /etc/nginx/conf.d/${DOMAIN}.conf<<-EOF
+            cat > ${NGINX_CONF_PATH}${DOMAIN}.conf<<-EOF
 server {
     listen 80;
     listen [::]:80;
@@ -602,7 +629,7 @@ EOF
             # VLESS+TCP+TLS
             # VLESS+TCP+XTLS
             # trojan
-            cat > /etc/nginx/conf.d/${DOMAIN}.conf<<-EOF
+            cat > ${NGINX_CONF_PATH}${DOMAIN}.conf<<-EOF
 server {
     listen 80;
     listen [::]:80;
@@ -616,11 +643,6 @@ server {
 }
 EOF
         fi
-    fi
-
-    certbotpath=`which certbot`
-    if [[ "$certbotpath" != "" ]]; then
-        echo "0 3 1 */2 0 root systemctl stop nginx ; ${certbotpath} renew; systemctl restart nginx" >> /etc/crontab
     fi
 }
 
@@ -750,7 +772,7 @@ installV2ray() {
         exit 1
     }
 
-    cat >/etc/systemd/system/v2ray.service<<-EOF
+    cat >$SERVICE_FILE<<-EOF
 [Unit]
 Description=V2ray Service
 Documentation=https://hijk.art
@@ -1193,7 +1215,11 @@ install() {
 
     $PMT clean all
     echo $CMD_UPGRADE | bash
-    $CMD_INSTALL wget net-tools unzip vim
+    $CMD_INSTALL wget vim unzip tar gcc openssl
+    $CMD_INSTALL net-tools
+    if [[ "$PMT" = "apt" ]]; then
+        $CMD_INSTALL libssl-dev g++
+    fi
     res=`which unzip`
     if [[ $? -ne 0 ]]; then
         colorEcho $RED " unzip安装失败，请检查网络"
@@ -1275,22 +1301,23 @@ uninstall() {
         
         stop
         systemctl disable v2ray
-        rm -rf /etc/systemd/system/v2ray.service
-        rm -rf /etc/systemd/system/multi-user.target.wants/v2ray.service
+        rm -rf $SERVICE_FILE
         rm -rf /etc/v2ray
         rm -rf /usr/bin/v2ray
-        
-        systemctl disable nginx
-        $CMD_REMOVE nginx
-        if [[ "$PMT" = "apt" ]]; then
-            $CMD_REMOVE nginx-common
-        fi
-        rm -rf /etc/nginx/nginx.conf
-        if [[ -f /etc/nginx/nginx.conf.bak ]]; then
-            mv /etc/nginx/nginx.conf.bak /etc/nginx/nginx.conf
+
+        if [[ "$BT" = "false" ]]; then
+            systemctl disable nginx
+            $CMD_REMOVE nginx
+            if [[ "$PMT" = "apt" ]]; then
+                $CMD_REMOVE nginx-common
+            fi
+            rm -rf /etc/nginx/nginx.conf
+            if [[ -f /etc/nginx/nginx.conf.bak ]]; then
+                mv /etc/nginx/nginx.conf.bak /etc/nginx/nginx.conf
+            fi
         fi
         if [[ "$domain" != "" ]]; then
-            rm -rf /etc/nginx/conf.d/${domain}.conf
+            rm -rf $NGINX_CONF_PATH${domain}.conf
         fi
         colorEcho $GREEN " V2ray卸载成功"
     fi
@@ -1302,7 +1329,8 @@ start() {
         colorEcho $RED " V2ray未安装，请先安装！"
         return
     fi
-    systemctl restart nginx
+    nginx -s stop
+    $START_NGINX
     systemctl restart v2ray
     sleep 2
     port=`grep port $CONFIG_FILE| head -n 1| cut -d: -f2| tr -d \",' '`
@@ -1315,7 +1343,7 @@ start() {
 }
 
 stop() {
-    systemctl stop nginx
+    nginx -s stop
     systemctl stop v2ray
     colorEcho $BLUE " V2ray停止成功"
 }
@@ -1363,7 +1391,7 @@ showInfo() {
         tls="true"
     fi
     if [[ "$ws" = "true" ]]; then
-        port=`cat /etc/nginx/conf.d/${domain}.conf | grep -i ssl | head -n1 | awk '{print $2}'`
+        port=`grep -i ssl $NGINX_CONF_PATH${domain}.conf| head -n1 | awk '{print $2}'`
     else
         port=`grep port $CONFIG_FILE | cut -d: -f2 | tr -d \",' '`
     fi
